* a monstrous commit, reorganizing files and code, renaming various functions, and stripping out some of AO’s ludicrously convoluted BS, to make overhauling 2D drawing more tractable (fades are currently disabled while we work on the rendering and it only gets as far as main menu); suffice to say it’ll be working when it’s working but at least its meandering code paths are starting to become short and clear; this is what happens when you tug gently on just one little thread: a hundred others pull outta the whole ball-o-yarn along with it; like attempting to shave a rabid yak while both rollerskating down the Andes, just gotta see the utter insanity of ever attempting it in the first place and hope that yak gets completely shaved before slamming into a boulder, plunging off a precipice, foaming at mouth also, etc, etc (luckily I’ve had some practice)

+++ Source_Files/RenderMain/textures.h
@@ -60,11 +60,11 @@ private:
 public:
 	bitmap_definition_buffer() {}
 	
-	explicit bitmap_definition_buffer(int row_count /*>= 1*/)
-		: buf(sizeof(bitmap_definition) + (/*assert(row_count >= 1), ? seriously, fuck off */ row_count - 1) * sizeof(pixel8*)) {}
+    // important: row_count MUST be >0
+	explicit bitmap_definition_buffer(int row_count) : buf(sizeof(bitmap_definition) + (row_count - 1) * sizeof(pixel8*)) {}
 	
 	bool empty() const { return buf.empty(); }
-	int row_count() const { return empty() ? 0 : 1 + (buf.size() - sizeof(bitmap_definition)) / sizeof(pixel8*); }
+	int row_count() const { return empty() ? 0 : 1 + ((int32_t)buf.size() - sizeof(bitmap_definition)) / sizeof(pixel8*); }
 	
 	bitmap_definition* get() { return empty() ? nullptr : reinterpret_cast<bitmap_definition*>(buf.data()); }
 	bitmap_definition const* get() const { return empty() ? nullptr : reinterpret_cast<const bitmap_definition*>(buf.data()); }
